# src-tauri/openapi.json
{
  "components": {
    "schemas": {
...
      "HttpMethod": {
        "description": "HTTP 方法（受管 HTTP 服务的自定义路由可指定）",
        "enum": [
          "GET",
          "POST",
          "PUT",
          "DELETE",
          "PATCH",
          "HEAD",
          "OPTIONS",
          "ANY"
        ],
        "type": "string"
      },
      "HttpRouteConfig": {
        "description": "HTTP 自定义路由配置（每个受管 HTTP 服务可配多条）",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "event": {
            "description": "固定事件名；若 None 且路径含 `{event}` 则取路径段；否则取末段",
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "method": {
            "$ref": "#/components/schemas/HttpMethod"
          },
          "path": {
            "description": "路径，支持 `{event}` 占位符（如 `/{event}`、`/order/{event}`）",
            "type": "string"
          },
          "routeType": {
            "$ref": "#/components/schemas/HttpRouteType"
          }
        },
        "type": "object"
      },
      "HttpRouteType": {
        "description": "HTTP 路由类型\n- Inbound：收消息（body 为 JSON），映射到 on_message\n- Stream：SSE 长连接，server→client 单向推送",
        "enum": [
          "inbound",
          "stream"
        ],
        "type": "string"
      },
...
      "LogLevel": {
        "description": "日志等级",
        "enum": [
          "DEBUG",
          "INFO",
          "WARN",
          "ERROR"
        ],
        "type": "string"
      },
...
      "MockMatchCondition": {
        "description": "头部/查询匹配条件",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "key": {
            "type": "string"
          },
          "matchKind": {
            "description": "匹配方式：exact / contains / regex / exists",
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "MockRule": {
        "description": "Mock 模拟规则",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "matchBody": {
            "description": "请求体匹配（包含子串；空表示不匹配 body）",
            "type": [
              "string",
              "null"
            ]
          },
          "matchHeaders": {
            "items": {
              "$ref": "#/components/schemas/MockMatchCondition"
            },
            "type": "array"
          },
          "matchQuery": {
            "items": {
              "$ref": "#/components/schemas/MockMatchCondition"
            },
            "type": "array"
          },
          "method": {
            "$ref": "#/components/schemas/HttpMethod",
            "description": "HTTP 方法（含 ANY）"
          },
          "name": {
            "type": "string"
          },
          "pathPattern": {
            "description": "路径模式：精确 `/users`、前缀 `/users/*`、参数 `/users/:id`",
            "type": "string"
          },
          "responseBody": {
            "type": "string"
          },
          "responseDelayMs": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "responseHeaders": {
            "items": {
              "$ref": "#/components/schemas/MockMatchCondition"
            },
            "type": "array"
          },
          "responseStatusCode": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "ProtocolType": {
        "description": "协议类型",
        "enum": [
          "websocket",
          "socket.io",
          "http"
        ],
        "type": "string"
      },
...
      "ServerConfig": {
        "description": "服务配置",
        "properties": {
          "autoStart": {
            "default": false,
            "type": "boolean"
          },
          "certPath": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "default": "",
            "type": "string"
          },
          "description": {
            "default": "",
            "type": "string"
          },
          "httpRoutes": {
            "default": [],
            "description": "HTTP 协议自定义路由（仅 protocol=http 生效；为空则用内置默认：\n`POST /{event}` 收消息 + `GET /stream` SSE 推送）",
            "items": {
              "$ref": "#/components/schemas/HttpRouteConfig"
            },
            "type": "array"
          },
          "id": {
            "default": "",
            "type": "string"
          },
          "ip": {
            "default": "0.0.0.0",
            "type": "string"
          },
          "keyPath": {
            "default": null,
            "type": [
              "string",
              "null"
            ]
          },
          "logLevel": {
            "default": "INFO",
            "oneOf": [
              {
                "$ref": "#/components/schemas/LogLevel"
              }
            ]
          },
          "mockDefaultDelayMs": {
            "default": 0,
            "description": "Mock 默认延迟（ms，未匹配规则时）",
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "mockDefaultResponseBody": {
            "default": "{\"message\":\"ok\"}",
            "description": "Mock 默认响应体（未匹配规则时返回）",
            "type": "string"
          },
          "mockDefaultStatusCode": {
            "default": 200,
            "description": "Mock 默认状态码（未匹配规则时返回）",
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "mockEnabled": {
            "default": false,
            "description": "是否在同一端口上同时提供 Mock HTTP 响应（统一路由模式）。\n启用后，非 WebSocket 升级的 HTTP 请求将由 Mock 引擎处理（规则匹配 → 模拟响应）。\nWebSocket 升级请求（含 `Upgrade: websocket` 头）仍由 Socket 传输层处理。\n两者在同一路由 fallback handler 中通过请求类型自动区分，互不干扰。",
            "type": "boolean"
          },
          "mockRules": {
            "default": [],
            "description": "Mock 规则列表（mock_enabled=true 时生效）；按顺序匹配，首个命中即返回",
            "items": {
              "$ref": "#/components/schemas/MockRule"
            },
            "type": "array"
          },
          "name": {
            "default": "",
            "type": "string"
          },
          "port": {
            "default": 0,
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "protocol": {
            "default": "websocket",
            "oneOf": [
              {
                "$ref": "#/components/schemas/ProtocolType"
              }
            ]
          },
          "updatedAt": {
            "default": "",
            "type": "string"
          },
          "wssEnabled": {
            "default": false,
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "ServerId": {
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Socket 服务管理平台后端 REST API（OpenAPI 3.1）。鉴权：Bearer token（`SSM_AUTH_ENABLED=1` 时生效）；`GET /api/auth/bootstrap` 回环自举获取 admin token。",
    "license": {
      "name": ""
    },
    "title": "NexSocket Studio · 管理面 API",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/server/add": {
      "post": {
        "operationId": "server_add",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": []
      }
    },
...
    "/api/server/remove": {
      "post": {
        "operationId": "server_remove",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerId"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": []
      }
    },
    "/api/server/restart": {
      "post": {
        "operationId": "server_restart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerId"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": []
      }
    },
...
    "/api/server/start": {
      "post": {
        "operationId": "server_start",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerId"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": []
      }
    },
...
    "/api/server/stop": {
      "post": {
        "operationId": "server_stop",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerId"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": []
      }
    },
...
    "/api/server/update": {
      "post": {
        "operationId": "server_update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": []
      }
    },
...
  },
  "servers": [
    {
      "description": "本地回环（默认绑定）",
      "url": "http://127.0.0.1:3080"
    }
  ]
}